refactor(sync): move the sync engine into the core crate
The app was the only place that could sync, so the engine lived in
src-tauri. The CLI is about to need it too, and writing a second copy
would fork the invariants that are held by comment alone -- never send
our own state, never rebuild state by rescanning after a download,
refuse a wholesale local deletion. Two copies drift, and the drift
deletes notes on every device. One engine, in core, is the only shape
that keeps that from happening.

Behaviour is unchanged: same wire format, same three CAS retries, same
refusal thresholds, and the same `kind` strings, which the TypeScript
error handling branches on (`lib/sync-status.ts`).

The HTTP-speaking half sits behind a new `sync-client` feature so the
Android JNI widget build and the MCP-only CLI keep paying nothing for
reqwest and keyring. The `reqwest::Client` is injected rather than built
in core: Android alone has to bypass the platform verifier
(`android_tls::sync_tls_config`, #175), and only the app knows that.
Both arms of `http_client()` stay in src-tauri.

src-tauri keeps the commands, `AppSyncState`, the events and the client
construction; auth.rs keeps the login window and the deep link path.

The 30 unit tests move with the code, assertions untouched.
